==> notebooks/0b_read_prism.ipynb <==
## Read PRSIM

calculate precipitation data for Midwest region (box)

import numpy as np
import pandas as pd
import xarray as xr
import dask.array as da
import glob as glob
import matplotlib.pyplot as plt
import cartopy.crs as ccrs
import cartopy.feature as cfeature

import geopandas as gpd
import regionmask
import cartopy.io.shapereader as sr

from project_utils import parameters as param

#### processed data files created in this notebook:

region_shp_file = "../processed_data/region.shp"
pr_mean_file = "../processed_data/prism_mean_"+str(param.time_period.start)+"_"+str(param.time_period.stop)+".nc"
region_mean_file = "../processed_data/region_mean_precip.csv"

#### Read PRISM data:

prism_path = "../input_data/PRISM/"
pr_prism_files = sorted(glob.glob(prism_path+"*.nc"))

pr_prism = xr.open_mfdataset(pr_prism_files, combine='nested', concat_dim='time').sel(
    lat = param.region_lat, lon = param.region_lon)

pr_prism['time'] = pd.to_datetime(pr_prism['time'].values, format='%Y%m%d')

#### Create Midwest region shapefile:

from shapely.geometry import Polygon

rx = np.interp(np.arange(0,100), [0, 25, 50, 75, 100], param.region_box_x) ## x coordinates
ry = np.interp(np.arange(0,100), [0, 25, 50, 75, 100], param.region_box_y) ## y coordinates

region_shp = gpd.GeoDataFrame(index=[0], crs='epsg:4326', 
                          geometry=[Polygon(zip(rx, ry))])

region_shp.to_file(filename=region_shp_file, driver="ESRI Shapefile")

region_mask = regionmask.mask_geopandas(region_shp, pr_prism)

#### Calculate mean precipitation (grid-level):

calc_climatology = True
## set to false to read already calculated data

if(calc_climatology):
    pr_mean = pr_prism.sel(time = param.time_period).mean(dim = "time")
    pr_mean.to_netcdf(pr_mean_file)

pr_mean = xr.open_dataset(pr_mean_file)

#### Regional precipitation time-series

region_mean = pr_prism.sel(lon = param.region_lon, lat = param.region_lat).mean(dim = ['lat', 'lon'])
region_mean = region_mean.to_dataframe()

region_mean.to_csv(region_mean_file)
