==> notebooks/1_train_cnn.ipynb <==
## Training the Convolutional Neural Network
This notebooks includes code to initialize and train the CNN. Output from this notebook includes: 
* Training history
* Trained weights of the CNN
* Class predictions for each day 

import numpy as np
import tensorflow as tf
import random

#### Set up reproducible conditions 
Set seed to generate reproducible results. 

**NOTE 1**: PYTHONHASHSEED=0 specified in kernel.json file (envs/share/jupyter/kernels/python3/kernel.json) (see https://stackoverflow.com/questions/58067359/is-there-a-way-to-set-pythonhashseed-for-a-jupyter-notebook-session/61953451#61953451). This variable must be set before kernel is launched. 

**NOTE 2**: current results are based on training on one CPU. Training on multiple CPUs will give very slightly different trained weights (but almost identical). 


np.random.seed(101)
random.seed(201)
tf.random.set_seed(333)

session_conf = tf.compat.v1.ConfigProto(device_count={'CPU': 1})
sess = tf.compat.v1.Session(config=session_conf)

import importlib
import pandas as pd
import matplotlib.pyplot as plt
import tensorflow.keras as keras

## project_utils are functions/variables defined by me
from project_utils import read_utils as read
from project_utils import utils as util
from project_utils import parameters as param
from project_utils import model_utils as mu

importlib.reload(read)
importlib.reload(util)
importlib.reload(mu)
importlib.reload(param)

#### load data and train model

x_dat = read.get_hgt_slp_input()

y_dat = util.get_precip_classes(pd.read_csv("../processed_data/region_mean_precip.csv")["prcp"], 
                               q = [0.95])

y_dat_onehot = util.onehot(y_dat)

ind = np.arange(len(y_dat))

from sklearn.model_selection import train_test_split

x_train, x_test, y_train, y_test, ind_train, ind_test = train_test_split(x_dat, y_dat_onehot, ind, 
                                                                         test_size=0.25, random_state=42, 
                                                                         shuffle = True, stratify = y_dat)

class_weights = util.class_weights(y_train[:,1].astype('int'))
class_weights

model = mu.build_model(activity_reg = 0.001, dropout_rate = 0.2)

model.load_weights("../processed_data/initial_weights.h5")

model.summary()

callback = tf.keras.callbacks.EarlyStopping(monitor='val_loss', patience=10, 
                                           restore_best_weights = True)

history = model.fit(x_train, y_train, 
          batch_size = 2048, 
          epochs = 400, 
          class_weight = class_weights, 
          validation_data = (x_test, y_test),          
          callbacks = [callback])

### model.evaluate(x_dat, y_dat_onehot)

model.save_weights("../processed_data/trained_weights.h5")

### Save history

hist_df = pd.DataFrame(history.history) 

hist_csv_file = '../processed_data/training_history.csv'
with open(hist_csv_file, mode='w') as f:
    hist_df.to_csv(f)

### Save predictions

class_predictions = model.predict(x_dat)

predict_df = pd.DataFrame(class_predictions)
predict_df = predict_df.rename(columns = {0: 'prob_0', 1: 'prob_1'})
predict_df['predicted_class'] = np.argmax(class_predictions, axis = 1)
predict_df['set'] = 'train'
predict_df.loc[ind_test, 'set'] = 'test'
predict_df['date'] = param.dates.values
predict_df['true_y'] = y_dat

predict_df.to_csv("../processed_data/predicted_class_data.csv", index = False)

### Model with geopotential height only

model_hgt = mu.build_model(activity_reg = 0.001, dropout_rate = 0.2, 
                                  input_channels = 1)

model_hgt.load_weights("../processed_data/initial_weights_hgt_only.h5")

x_dat_hgt = read.get_hgt_input()
x_train_hgt = x_dat_hgt[ind_train]
x_test_hgt = x_dat_hgt[ind_test]

history_hgt = model_hgt.fit(x_train_hgt, y_train, 
          batch_size = 2048, 
          epochs = 400, 
          class_weight = class_weights, 
          validation_data = (x_test_hgt, y_test),          
          callbacks = [callback])

model_hgt.evaluate(x_dat_hgt, y_dat_onehot)

model_hgt.save_weights("../processed_data/trained_weights_hgt_only.h5")

hist_df_hgt = pd.DataFrame(history_hgt.history) 

hist_csv_file = '../processed_data/training_history_hgt_only.csv'
with open(hist_csv_file, mode='w') as f:
    hist_df_hgt.to_csv(f)

class_predictions_hgt = model_hgt.predict(x_dat_hgt)

predict_df_hgt = pd.DataFrame(class_predictions_hgt)
predict_df_hgt = predict_df_hgt.rename(columns = {0: 'prob_0', 1: 'prob_1'})
predict_df_hgt['predicted_class'] = np.argmax(class_predictions, axis = 1)
predict_df_hgt['set'] = 'train'
predict_df_hgt.loc[ind_test, 'set'] = 'test'
predict_df_hgt['date'] = param.dates.values
predict_df_hgt['true_y'] = y_dat

predict_df_hgt.to_csv("../processed_data/predicted_class_data_hgt_only.csv", index = False)
