==> notebooks/2_lrp.ipynb <==
## Layer-wise Relevance Propagation

Implement LRP using iNNvestigate package: https://innvestigate.readthedocs.io/en/latest/index.html. iNNvestigate currently works with keras models and tensorflow 1.X. In this notebook, the sequential model is recreated in keras, and the trained weights are loaded. 

link to older Keras documentation: https://faroit.com/keras-docs/2.0.2/metrics/#binary_accuracy

import numpy as np
import pandas as pd
import xarray as xr

import keras

import innvestigate

import importlib

## project_utils are functions/variables defined by me
from project_utils import parameters as param
from project_utils import utils as util
from project_utils import read_utils as read

importlib.reload(read)
importlib.reload(util)
importlib.reload(param)

### Build keras model and load pre-trained weights

## slightly different syntax to build model in TF1 (comments note differences)
def build_model_keras(lr = .0004, conv_filters = 16, dense_neurons = 16, 
                             dense_layers = 1, 
                activity_reg = 0.001, dropout_rate = 0.2, input_channels = 2):

    model = keras.models.Sequential() 
    model.add(keras.layers.InputLayer(input_shape=(15, 35, input_channels))) ## define input shape
    ## InputLayer instead of Input, input_shape instead of shape
    
    model.add(keras.layers.Conv2D(conv_filters, 
                            (3,3), activity_regularizer=keras.regularizers.l2(0.01))) 
    model.add(keras.layers.Activation('relu')) 
    model.add(keras.layers.MaxPooling2D((2,2)))
                            
    model.add(keras.layers.Conv2D(conv_filters, (3,3), activity_regularizer=keras.regularizers.l2(activity_reg))) 
    model.add(keras.layers.Activation('relu')) 
    model.add(keras.layers.MaxPooling2D((2,2))) 
                            
    model.add(keras.layers.Flatten()) 
    for i in range(dense_layers):
        model.add(keras.layers.Dense(dense_neurons, activity_regularizer=keras.regularizers.l2(activity_reg)))
        model.add(keras.layers.Activation('relu'))
     
    model.add(keras.layers.Dense(2, activation='softmax')) 
    
    model.compile(loss=keras.losses.categorical_crossentropy, ## instead of CategoricalCrossentropy
                  optimizer=keras.optimizers.Adam(lr = lr), ## lr instead of learning_rate
                  metrics=['categorical_accuracy'])
    return(model)

model = build_model_keras()

model.load_weights("../processed_data/trained_weights.h5")

x_dat = read.get_hgt_slp_input()
y_dat = util.get_precip_classes(pd.read_csv("../processed_data/region_mean_precip.csv")["prcp"], 
                               q = [0.95])

y_dat_onehot = util.onehot(y_dat)

ind = np.arange(len(y_dat))

from sklearn.model_selection import train_test_split

x_train, x_test, y_train, y_test, ind_train, ind_test = train_test_split(x_dat, y_dat, ind, test_size=0.25, random_state=42, 
                                                   shuffle = True, stratify = y_dat)

#### compare predictions to TF2

model.evaluate(x_dat, y_dat_onehot)

^ same accuracy as tf2 model (very small error <0.0000001)

predict_df = pd.read_csv("../processed_data/predicted_class_data.csv")

test_predictions = predict_df.loc[predict_df.set == 'test']

test_predictions_tf1 = np.argmax(model.predict(x_dat)[np.sort(ind_test)], axis = 1)

np.sum(test_predictions_tf1 != test_predictions.predicted_class)

^ all predictions match. (Also checked that probabilities match)

## LRP

see Montavon et al. (https://www.sciencedirect.com/science/article/pii/S1051200417302385)

LRP algorithm using Alpha-Beta rule - Alpha1Beta0 only tracks positive relevance 

model_strip = innvestigate.utils.model_wo_softmax(model)

## create analyzer object 
lrp_analyzerA1B0 = innvestigate.analyzer.relevance_based.relevance_analyzer.LRPAlpha1Beta0(model_strip)

### save relevance data for all days

rel_all = lrp_analyzerA1B0.analyze(x_dat)

rel_data = xr.Dataset(
    data_vars=dict(
        rel_hgt=(["time", "lat", "lon"], rel_all[:,:,:,0]), 
        rel_slp=(["time", "lat", "lon"], rel_all[:,:,:,1]),
    ),
    coords=dict(
        time=predict_df.date.values, 
        lat = param.lats, 
        lon = param.lons),
)

rel_data

rel_data.to_netcdf("../processed_data/lrp_data_all_days.nc")
